+++ app/components/TxHistory/TxHistoryRow/RegularTxRow.js
