+++ app/components/TxHistory/TxHistoryRow/StakeTxRow.js
