+++ app/components/TxHistory/TxHistoryRow/Status.js
@@ -1,5 +1,5 @@
 import { FormattedMessage as T } from "react-intl";
-import { tsToDate } from "../../helpers/dateFormat";
+import { tsToDate } from "helpers/dateFormat";
 import "style/TxHistory.less";
 
 // TODO: use a global component for these indicators
