+++ app/components/TxHistory/TxHistoryRow/index.js
