+++ app/components/shared/index.js
@@ -1,11 +1,9 @@
+export { default as Balance } from "./Balance";
 export { default as CopyToClipboard } from "./CopyToClipboard";
 export { default as RouteTransition } from "./RouteTransition";
 export { default as LinkToAccounts } from "./LinkToAccounts";
+export { default as ShowError } from "./ShowError";
 export { default as TabbedHeader } from "./TabbedHeader";
 export { default as TabbedPage } from "./TabbedPage";
 export { default as Tooltip } from "./Tooltip";
-
 export { default as TransitionMotionWrapper } from "./TransitionMotionWrapper";
-export { default as Balance } from "./Balance";
-export { default as TxHistory } from "./TxHistory";
-export { default as ShowError } from "./ShowError";
