+++ app/components/views/TicketsPage/PurchaseTab/PurchaseTickets/index.js
