+++ app/components/views/TicketsPage/PurchaseTab/StakeInfo/Display.js
