+++ app/components/views/TicketsPage/PurchaseTab/StakeInfo/StakeInfoDetails.js
