+++ app/components/views/TicketsPage/PurchaseTab/StakeInfo/StakeInfoRow.js
