+++ app/components/views/TicketsPage/PurchaseTab/StakeInfo/index.js
