+++ app/components/views/TicketsPage/PurchaseTab/StakePools/AddForm.js
