+++ app/components/views/TicketsPage/PurchaseTab/StakePools/List.js
