+++ app/components/views/TicketsPage/PurchaseTab/StakePools/index.js
