+++ app/components/views/TicketsPage/PurchaseTab/TicketAutoBuyer/index.js
