+++ app/components/views/TransactionsPage/HistoryTab/Page.js
@@ -1,6 +1,7 @@
 import EyeFilterMenu from "EyeFilterMenu";
 import { FormattedMessage as T } from "react-intl";
-import { Tooltip, TxHistory } from "shared";
+import { Tooltip } from "shared";
+import TxHistory from "TxHistory";
 import { LoadingMoreTransactionsIndicator, NoMoreTransactionsIndicator } from "indicators";
 import InfiniteScroll from "react-infinite-scroller";
 import "style/HistoryPage.less";
